+++ scripts/macros/2024/spells/massCureWounds.js
@@ -1,0 +1,6 @@
+import {massCureWounds as massCureWoundsLegacy} from '../../../legacyMacros.js';
+export let massCureWounds = {
+    ...massCureWoundsLegacy,
+    version: '1.2.24',
+    rules: 'modern'
+};