+++ scripts/macros/2024/spells/trueStrike.js
@@ -40,6 +40,7 @@ async function use({workflow}) {
 export let trueStrike = {
     name: 'True Strike',
     version: '1.2.24',
+    rules: 'modern',
     midi: {
         item: [
             {